Control.Concurrent.Chan.Lifted

Description

This is a wrapped version of Control.Concurrent.Chan with types generalised from IO to all monads in MonadBase.

unGetChan and isEmptyChan are deprecated in base, therefore they are not included here. Use TVar instead.

Documentation

data Chan a

Chan is an abstract type representing an unbounded FIFO channel.

newChan :: MonadBase IO m => m (Chan a)

Generalized version of newChan.

writeChan :: MonadBase IO m => Chan a -> a -> m ()

Generalized version of writeChan.

readChan :: MonadBase IO m => Chan a -> m a

Generalized version of readChan.

dupChan :: MonadBase IO m => Chan a -> m (Chan a)

Generalized version of dupChan.

getChanContents :: MonadBase IO m => Chan a -> m [a]

Generalized version of getChanContents.

writeList2Chan :: MonadBase IO m => Chan a -> [a] -> m ()

Generalized version of writeList2Chan.