Control.Concurrent.QSem.Lifted

Description

This is a wrapped version of Control.Concurrent.QSem with types generalised from IO to all monads in MonadBase.

Documentation

data QSem

QSem is a quantity semaphore in which the resource is acquired and released in units of one. It provides guaranteed FIFO ordering for satisfying blocked waitQSem calls.

The pattern

  bracket_ waitQSem signalQSem (...)

is safe; it never loses a unit of the resource.

newQSem :: MonadBase IO m => Int -> m QSem

Generalized version of newQSem.

waitQSem :: MonadBase IO m => QSem -> m ()

Generalized version of waitQSem.

signalQSem :: MonadBase IO m => QSem -> m ()

Generalized version of signalQSem.