Control.Concurrent.QSemN.Lifted

Description

This is a wrapped version of Control.Concurrent.QSemN with types generalised from IO to all monads in MonadBase.

Documentation

data QSemN

A QSemN is a quantity semaphore, in which the available "quantity" may be signalled or waited for in arbitrary amounts.

newQSemN :: MonadBase IO m => Int -> m QSemN

Generalized version of newQSemN.

waitQSemN :: MonadBase IO m => QSemN -> Int -> m ()

Generalized version of waitQSemN.

signalQSemN :: MonadBase IO m => QSemN -> Int -> m ()

Generalized version of signalQSemN.