Foreign.Marshal.Utils.Lifted

Description

This is a wrapped version of Foreign.Marshal.Utils with types generalized from IO to all monads in either MonadBase or MonadBaseControl.

Documentation

with

Arguments

:: (MonadBaseControl IO m, Storable a)

=> a
value to be poked

-> (Ptr a -> m b)
computation to run

-> m b

Generalized version of with.

Note, when the given function throws an exception any monadic side effects in m will be discarded.