System.Timeout.Lifted

Description

Attach a timeout event to monadic computations which are instances of MonadBaseControl.

Documentation

timeout :: MonadBaseControl IO m => Int -> m a -> m (Maybe a)

Generalized version of timeout.

Note that when the given computation times out any side effects of m are discarded. When the computation completes within the given time the side-effects are restored on return.